extract generic StyleManager
