+++ src/style/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Created by Samuel Gratzl on 26.09.2017.
+ */
+
+export {IColumn} from './IColumn';
+export {default as GridStyleManager, setColumn, TEMPLATE, setTemplate} from './GridStyleManager';
+export {default as StyleManager} from './StyleManager';
